requirements.txt
# Framework web para el servidor
Flask==2.3.3

# Criptografía - AES y funciones seguras
pycryptodome==3.18.0

# Requests HTTP para el cliente
requests==2.31.0

# Criptografía adicional (opcional pero recomendado)
cryptography==41.0.4

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# INSTALACIÓN
# pip install -r requirements.txt
# 
